comment shrs_line
