+++ crates/shrs_line/src/cursor.rs
@@ -4,6 +4,7 @@ use crossterm::cursor::SetCursorStyle;
 
 #[derive(Clone)]
 pub struct CursorStyle {
+    /// The cursor style used by crossterm to draw the cursor
     pub style: SetCursorStyle,
 }
