+++ crates/shrs_line/src/highlight.rs
@@ -1,3 +1,5 @@
+//! Syntax highlighting
+
 use std::{collections::HashMap, usize};
 
 use crossterm::style::{Color, ContentStyle};
