+++ crates/shrs_line/src/painter.rs
@@ -1,3 +1,5 @@
+//! Internal renderer
+
 use std::{
     collections::HashMap,
     fmt::Display,
