+++ src/lib.rs
@@ -59,7 +59,7 @@ pub mod prelude {
 
     pub use shrs_core::prelude::*;
     pub use shrs_lang::PosixLang;
-    pub use shrs_line::{completion::*, *};
+    pub use shrs_line::prelude::*;
 
     pub use crate::{anyhow, crossterm, crossterm::*, plugin::*, shell::*};
 }
